requirements.txt
# Core ML packages
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# Data processing
numpy>=1.21.0
opencv-python>=4.5.0
pillow>=8.0.0
matplotlib>=3.3.0
seaborn>=0.11.0
pandas>=1.3.0

# Utilities
pyyaml>=6.0
tqdm>=4.60.0
requests>=2.25.0

# Jupyter (if using notebooks)
jupyter>=1.0.0
ipywidgets>=7.6.0

# Optional GPU support (uncomment if using NVIDIA GPU)
# torch-torchvision-cu118  # For CUDA 11.8
# torch-torchvision-cu121  # For CUDA 12.1
